+++ CHANGELOG.md
@@ -22,13 +22,17 @@
 - `ModelText`
 - `Solid.Intersects(Plane p, out List<Polygon> result)`
 - `Vector3.IsUnitized()`
-- `Transform.Inverse()`
+- `Transform.Inverted()`
+- `AdaptiveGrid`
+- `Line.Intersects(BBox3 box, out List<Vector3> results, bool infinite = false)`
+- `Vector3.AreCoplanar(Vector3 a, Vector3 b, Vector3 c, Vector3 d)`
 
 ### Changed
 
 ### Fixed
 
 - Deduplicate catalog names during code generation.
+- Fixed an issue where GLTFs would occasionally be generated with incorrect vertex normals.
 
 ## 0.9.2